+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 - |Enhancement| Metrics built with the :class:`~empulse.metrics.MaxProfit`
   strategy can now handle instance-dependent costs.
+- |Fix| Fix :class:`~empulse.models.CSTreeClassifier` not properly training when costs were negative.
 
 `0.10.4`_ (20-09-2025)
 ======================